refactor: Reorganize and improve naming
