+++ src/app/[locale]/(main)/dashboard/[dashboardNameAndId]/page.tsx
@@ -1,6 +1,6 @@
 import { concatenateNameAndId, safeStringCompare } from '@/utils/stringUtils';
 
-import DetailsPageControls from '@/components/details-page/DetailsPageControls';
+import DashboardResourceControls from '@/components/dashboard-resource/DashboardResourceControls';
 import ErrorComponent from '@/components/error-handling/ErrorComponent';
 import PageWrapper from '@/components/layout/PageWrapper';
 import { getValidatedConfiguration } from '@/schemas/validate';
@@ -20,7 +20,7 @@ export default async function Page({ params: { dashboardNameAndId } }: { params:
 
   return (
     <PageWrapper title={dashboard.name} description={dashboard.description}>
-      <DetailsPageControls type="dashboard" element={dashboard} />
+      <DashboardResourceControls type="dashboard" element={dashboard} />
       <>Some dashboard content</>
     </PageWrapper>
   );
