+++ src/components/dashboard-resource/ResourceDetails.tsx
@@ -4,12 +4,14 @@ import EmbeddedViewer from '../visualization/EmbeddedViewer';
 import Visualization from '../visualization/Visualization';
 import dynamic from 'next/dynamic';
 
-const DetailsPageControls = dynamic(() => import('../details-page/DetailsPageControls'), { ssr: false });
+const DashboardResourceControls = dynamic(() => import('./DashboardResourceControls'), {
+  ssr: false,
+});
 
 export default function ResourceDetails({ resource }: { resource: Resource }) {
   return (
-    <div>
-      <DetailsPageControls type="resource" element={resource} />
+    <>
+      <DashboardResourceControls type="resource" element={resource} />
       <div
         className={`d-flex flex-column flex-grow-1 ${resource.type === 'GeoJSON' ? 'border border-secondary' : ''}`}
         style={{
@@ -23,6 +25,6 @@ export default function ResourceDetails({ resource }: { resource: Resource }) {
           <Visualization resource={resource as JSONResource | GeoJSONResource} />
         )}
       </div>
-    </div>
+    </>
   );
 }
